+++ ms-pagamento/src/main/java/com/senac/torreverde/ms_pagamento/dtos/PedidoDTO.java
@@ -1,6 +1,7 @@
 package com.senac.torreverde.ms_pagamento.dtos;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.List;
 
 // Classe DTO para recebimento do evento de criar Pedido via Rabbitmq
@@ -8,13 +9,13 @@ public class PedidoDTO implements Serializable {
 
     private int pedidoId;
     private Integer usuarioId;
-    private Double pedidoValorTotal;
+    private BigDecimal pedidoValorTotal;
     private List<PedidoItemDTO> itens;
 
     public PedidoDTO() {
     }
 
-    public PedidoDTO(int pedidoId, Integer usuarioId, Double pedidoValorTotal, List<PedidoItemDTO> itens) {
+    public PedidoDTO(int pedidoId, Integer usuarioId, BigDecimal pedidoValorTotal, List<PedidoItemDTO> itens) {
         this.pedidoId = pedidoId;
         this.usuarioId = usuarioId;
         this.pedidoValorTotal = pedidoValorTotal;
@@ -37,11 +38,11 @@ public class PedidoDTO implements Serializable {
         this.usuarioId = usuarioId;
     }
 
-    public Double getPedidoValorTotal() {
+    public BigDecimal getPedidoValorTotal() {
         return pedidoValorTotal;
     }
 
-    public void setPedidoValorTotal(Double pedidoValorTotal) {
+    public void setPedidoValorTotal(BigDecimal pedidoValorTotal) {
         this.pedidoValorTotal = pedidoValorTotal;
     }
