+++ ms-pagamento/src/main/java/com/senac/torreverde/ms_pagamento/dtos/PedidoItemDTO.java
@@ -1,17 +1,19 @@
 package com.senac.torreverde.ms_pagamento.dtos;
 
+import java.math.BigDecimal;
+
 // Classe DTO para recebimento do evento de criar PedidoItem via Rabbitmq
 public class PedidoItemDTO {
 
     private Integer estoqueId;
     private Integer quantidade;
-    private Double precoUnitario;
-    private Double subTotal;
+    private BigDecimal precoUnitario;
+    private BigDecimal subTotal;
 
     public PedidoItemDTO() {
     }
 
-    public PedidoItemDTO(Integer estoqueId, Integer quantidade, Double precoUnitario, Double subTotal) {
+    public PedidoItemDTO(Integer estoqueId, Integer quantidade, BigDecimal precoUnitario, BigDecimal subTotal) {
         this.estoqueId = estoqueId;
         this.quantidade = quantidade;
         this.precoUnitario = precoUnitario;
@@ -34,19 +36,19 @@ public class PedidoItemDTO {
         this.quantidade = quantidade;
     }
 
-    public Double getPrecoUnitario() {
+    public BigDecimal getPrecoUnitario() {
         return precoUnitario;
     }
 
-    public void setPrecoUnitario(Double precoUnitario) {
+    public void setPrecoUnitario(BigDecimal precoUnitario) {
         this.precoUnitario = precoUnitario;
     }
 
-    public Double getSubTotal() {
+    public BigDecimal getSubTotal() {
         return subTotal;
     }
 
-    public void setSubTotal(Double subTotal) {
+    public void setSubTotal(BigDecimal subTotal) {
         this.subTotal = subTotal;
     }
 }
